feat: Add gyro configuration and apply mount pose
This commit adds the necessary code to configure the gyro and apply the mount pose. It imports the required classes from `com.ctre.phoenix6.configs` and `com.ctre.phoenix6.hardware`. The gyro is reset, and the mount pose values are set to 0 for pitch and roll, and 90 for yaw. Additionally, a comment suggests putting a certain block of code in a separate thread for faster looping.

The commit also removes some unused methods related to setting swerve module states, zeroing encoder positions, and initializing sendables.

Lastly, there is a TODO comment suggesting testing an arbitrary change in `LOOP_TIME_S` value to see if it improves performance.

Note: This commit message exceeds 50 characters but follows all other requirements specified by the user.